+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 - Disable RunningSpanStore by default unless the z-pages contrib is linked and initialized. This may
 break behaviors for applications that rely on this to be always enabled.
 - Provide a `Deadline` option to Stackdriver Trace exporter. Default value is 10 seconds if it's not set.
-- Provide a `Deadline` option to Stackdriver Stats exporter, and a `MetricServiceStub` option so that
-advanced users can use a custom Stackdriver Monitoring client to make RPCs.
+- Provide a `Deadline` option to Stackdriver Stats exporter. Default value is 10 seconds.
+Also provide a `MetricServiceStub` option so that advanced users can use a custom Stackdriver
+Monitoring client to make RPCs.
 
 ## 0.21.0 - 2019-04-30
 - Add HTTP text format serializer to Tag propagation component.